+++ modules/post-processing/lane-detection/stack.py
@@ -18,6 +18,7 @@ class LaneDetection(Stack):
         scope: Construct,
         id: str,
         *,
+        project_name: str,
         deployment_name: str,
         module_name: str,
         s3_access_policy: str,
@@ -34,7 +35,7 @@ class LaneDetection(Stack):
             value="aws",
         )
 
-        dep_mod = f"addf-{deployment_name}-{module_name}"
+        dep_mod = f"{project_name}-{deployment_name}-{module_name}"
 
         self.repository_name = dep_mod
         repo = ecr.Repository(self, id=self.repository_name, repository_name=self.repository_name)
@@ -44,7 +45,7 @@ class LaneDetection(Stack):
             iam.PolicyStatement(
                 actions=["dynamodb:*"],
                 effect=iam.Effect.ALLOW,
-                resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/addf*"],
+                resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/{project_name}*"],
             ),
             iam.PolicyStatement(
                 actions=["ecr:*"],
@@ -54,7 +55,7 @@ class LaneDetection(Stack):
             iam.PolicyStatement(
                 actions=["s3:GetObject", "s3:GetObjectAcl", "s3:ListBucket"],
                 effect=iam.Effect.ALLOW,
-                resources=["arn:aws:s3:::addf-*", "arn:aws:s3:::addf-*/*"],
+                resources=[f"arn:aws:s3:::{project_name}-*", f"arn:aws:s3:::{project_name}-*/*"],
             ),
         ]
         dag_document = iam.PolicyDocument(statements=policy_statements)